+++ src/main/java/com/siftscience/model/AbuseScore.java
@@ -8,6 +8,9 @@ import java.util.List;
 public class AbuseScore {
     @Expose @SerializedName("score") private Double score;
     @Expose @SerializedName("reasons") private List<Reason> reasons;
+    @Expose @SerializedName("status") private int status;
+    @Expose @SerializedName("error_message") private String errorMessage;
+    @Expose @SerializedName("time") private long time;
 
     public Double getScore() {
         return score;
@@ -24,6 +27,33 @@ public class AbuseScore {
 
     public AbuseScore setReasons(List<Reason> reasons) {
         this.reasons = reasons;
+        return this;
+    }
+
+    public int getStatus() {
+        return status;
+    }
+
+    public AbuseScore setStatus(int status) {
+        this.status = status;
+        return this;
+    }
+
+    public String getErrorMessage() {
+        return errorMessage;
+    }
+
+    public AbuseScore setErrorMessage(String errorMessage) {
+        this.errorMessage = errorMessage;
+        return this;
+    }
+
+    public long getTime() {
+        return time;
+    }
+
+    public AbuseScore setTime(long time) {
+        this.time = time;
         return this;
     }
 }
